Remove CDO from optional data when it's populated (#9073)
## Describe your changes
Improves some semantics for this onboarding requirement. As is, a CDO
moves from `required_attributes` to `populated_attributes` once it is
collected. But strangely, CDOs stay in `optional_attributes` even when
they are collected. So it's hard for the frontend to actually tell if an
optional CDO is collected - it would actually decrypt data from the
backend in order to see if it was populated...

This lets us simplify a good amount of frontend code by simply treating
`optional_attributes` as data that needs to be collected, but also may
be skipped.

## Caveats

I haven't been a fan of the API for optional data for some time: right
now, a CDO exists in `optional_data` but not `must_collect_data`. This
requires a lot of locations to chain together `must_collect_data` and
`optional_data`, which is super error prone. I'd instead prefer if
`optional_data` were the subset of `must_collect_data` that is optional.
But I'm afraid this ship has sailed... a migration of this scale would
not really be practical

## Test Plan
Integration test 


<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->