[identify] Add unit test that we send x-fp-is-components-sdk to backend (#9044)
## Describe your changes
In #9022 and #9039, we started sending some new flags to the backend that indicate whether data is (1) added via bootstrap or (2) added via the components SDK. This PR just adds a test that we send these flags correctly.

To do so, I had to add a new method of asserting that the body and headers of an HTTP request match what we expect - I tried a few mechanisms but couldn't find a good way that gave a reasonable error message upon failure. I ended up reusing `jest.fn()` here so we can use our normal `expect(...).toEqual(...)` utils, but it takes a little complexity to actually extract the contents of the HTTP request - lmk what you think of this method :)

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
they are tests!