[FE CI] Fix E2E test flakes from demo page initialization race condition (#9095)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

This logic from the demo page didn't wait until the router was ready.
Instead, it directly tried to grab the public key or defaulted to the
one from .env file. I speculate that this was the reason behind most of
the flakes we were seeing. Since the e2e testing framework acts very
quickly we launch with incorrect ob config and tests broke.

Solution:
- Wait for router to be ready and update the ob config and other vars
correctly
- Explicitly show a loading state on the demo page before the router is
initialized and the ob key is read

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->