Simplify KYC machine transitions (#9064)
## Describe your changes
I stumbled upon some of this complex logic when working on the KYC
machine and wanted to make it more accommodating for those who come
after me :) I think a few small changes to the KYC machine over time
have enabled us to drastically simplify and de-dupe logic, but we didn't
make the larger refactor alongside these small changes.

A few key observations that allowed us to simplify this logic:
- forward transitions would use `isMissing(currentData,
ignoreCollectedData = true)`, which was basically identical to
`isMissing(initialData)`. I think initial data was added later and we
didn’t simplify state machine transitions when we added it
- with this knowledge that state machine transitions are mostly a
function of isMissing(initialData) , we can statically compute the set
of pages that will be shown when the machine is initialized instead of
computing them in each transition. this simplified a lot of logic that
would compute which step of the onboarding process we were on
- the transitions between pages were copy-pasted, but we can instead
write one generic set of transitions that dictate flow between states
instead of custom transition rules for each state

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Existing test coverage - this didn't break any tests in non-trivial
ways.

Lots of manual testing:

*Normal flow collecting everything*



https://github.com/onefootprint/monorepo/assets/21066341/faa9a9c8-59cf-4d12-80f5-efdf78bc8cc4


*Entire user data already exists, no need to collect anything*



https://github.com/onefootprint/monorepo/assets/21066341/73eb7531-6ebf-47b4-b51c-194157cfe54a


*Partial user data already exists*


https://github.com/onefootprint/monorepo/assets/21066341/49769bba-960a-4142-a482-fad33a789bb0