Start logging to work on fixing skip SSN bug (#9091)
## Describe your changes
Found another bug here - if you click "skip SSN" but already have an SSN vaulted, we won't actually clear it. This isn't necessarily the worst side effect since we're at least not losing information. But worth fixing in the future.

Would need to stop clearing empty values in `useUserData`, so i added some logging to it to see if we rely on that codepath for anything else.

Also added an error log if someone does try to skip SSN but they already have one, just to give us a gauge of how often people are hitting this bug

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->