Some updates to curp so it works better in sandbox and we vault data (#9107)
## Describe your changes
**Sandbox**
One motivation of this PR is to support curp in sandbox

Currently the way CURP works is we collect doc, then when making vendor calls, we retrieve the OcrDataKind::Curp from the most recent document and send it off to incode

However, in sandbox there's a few nuances:
* we don't necessarily send data to incode, so we don't get OCR
   * this could be mitigated by providing fixtures, but this seems more brittle and hidden than just handling in the manner in this PR
* additionally, we want to be aware of the fact we're in prod and we should error if we expect a CURP and don't find one

**Vaulting:**
CURP (and government validation) returns a lot of data that is helpful in a review. However, processing this data (with risk signals) or displaying this data (in the dashboard in the doc drawer or elsewhere) is a lot of work.

Instead, let's just vault the response and let Coba view the data alongside other OCR/document data and make their own decisions. We'll show this in the ocr data area in the drawer in a collapsable view (ahsan will help with this part)
## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->