Demo: remove ghost dependency (#9170)
## Describe your changes

In the past, we used to use `Ghost` to host some custom demos for some
leads. The idea was to upload some logos and texts, that Eli could use
to present to customers. However, we have been using Coast
(https://www.trycoast.com/) to do this. So there's no need to use Ghost
anymore. This will simplify our code

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->