make sure we don't have half neuro sessions (#9180)
## Describe your changes
Since neuro is not an immutable part of a playbook, we must toggle on (and off) the feature flag. The FE knows to load neuro js based on the feature flag being return in the `PublicOBC` which is called **before** identify/verify/we get_or_create an onboarding/workflow.

Neuro also is 2 stages:
* FE collection (where the BE provides an identifier that it uses to send the metrics collection)
* BE API call (after FE collection) where we fetch results from neuro using the same identifier ("salted" wf_id)

Lastly, feature flags can turn on and off at any time, so we expect that this will be in the middle of some user's workflow.

This means that we need to keep the user that is onboarding and the BE fetching data in sync across a few cases:
* users that already started when the flag turns on  -> we should not try and fetch Neuro ID (worst case is a BE error)
* users that already started when the flag turns off -> we theoretically could try and run neuro, but it's unclear if the FE captured the full workflow's events for a given WF. for example, did they exit and come back to the workflow? (worst case is weird signal from Neuro)

Instead, we can cache whether the workflow started with neuro enabled, and assert when we go to make the BE neuro call that the FF matches what the WF is expecting. As long as we aren't flipping the flag on and off randomly, we will avoid any partial instance of FE (not) collecting and the BE making a request for a neuro decision with unknown incomplete data.

This is also important in the PR above this one - we want to not generate both stytch AND neuro ID signals

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
I don't have a good way to test this. But the entire neuro call will not error closed, so we can debug  on the fly

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->