Fix issue where FileTooLarge errors cause client timeouts (#9176)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
In some cases, if the client uploads a file over the limit and the
server responds with a `FileTooLarge` error without reading the payload,
the client may hang until the client timeout is reached. I've only been
able to reproduce this locally -- the load balancers seem to have a
different behavior -- perhaps they buffer the payload internally.

The solution I'm applying in this PR is to close the connection upon
sending a `FileTooLarge` HTTP 413 error. This seems to work.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->