Ensure new list rules have compatible list kind and field type (#9151)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical details and possible side effects. -->
We want to make sure we can't create a rule that attempts to check membership of a value in a list that contains values of an incompatible type. e.g. Emails can only be matched against email lists and email domain lists.

## Caveats
<!--- Are there any caveats to this changeset? -->
This PR admits business phone numbers and also custom fields in rules eval -- I'm not 100% sure if those are something we want to support?

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->