Use DocumentRequestKind in WorkflowRequestConfig and WorkflowConfig (#9253)
## Describe your changes
We're starting to thread DocumentRequestConfig all the way through from the triggers API -> WorkflowRequest -> Workflow. We've had different representations of "document requirements" in each of these stages, and we can finally consolidate into one representation (that's also used on playbooks and DocumentRequests)!

This starts writing the newer forms of WorkflowRequestConfig and WorkflowConfig - in a future PR, we'll backfill all old workflow requests and workflows, and then we'll be able to drop these old variants

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->