Add review_status to identity document (#9280)
## Describe your changes
We'll use this to track when a human or a machine has reviewed a
document. All documents start in `requires_review` . they can transition
to `reviewed_by_machine` if we automatically review the document w
incode (or in the future, ocrulus). or, they can transition to
`reviewed_by_human` if a human explicitly reviews the document.

## Caveats

There are also some changes coming to the manual review model +
dashboard that help to make this more explicit


## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->