[FP-7499] Mark document as human-reviewed when clearing ManualReview for docs (#9304)
## Describe your changes
When a manual review is raised that specifically says documents need to be reviewed, reviewing the user will now explicitly mark the _document_ as reviewed.

## Caveats

This is pretty implicit - not my favorite. but gets the job done for now until we have better per-document status updating. maybe we can remove this implicit behavior in the future, or somehow make it more explicit in the UI

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->