Share step up logic between document and KYC workflows (#9306)
## Describe your changes
IIRC, this step-up logic was recently added to the document workflow. This just moves it to a shared location so it can be used for both the KYC and document workflows.

Consolidating logic like this brings us sliiiiightly closer to the one-workflow-kind future... still have a ways to go, but it at least makes it very clear that logic between workflows is identical :)

## Caveats

I think the KYB workflow could also use this new common logic too? maybe we just didn't add step-up logic to it

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->