[FP-7507] Support no rules executing for a Document playbook (#9344)
## Describe your changes
We want the default behavior to be to not execute any rules for document-only playbooks that only collect custom (/PoA/PoSsn) docs. We're going to make this ~slightly more explicit by allowing these playbooks to have no rules. When this is the case, we will maintain the same scoped vault status that was existing before.

## Caveats

I chose that if the first playbook a user onboards onto has no rules, we will default to Pass (similar to if no rule is matched on a playbook that has rules). In the PR on top of this, we'll put the user into fail if they have a document manual review

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->