Rely on ready-state instead of onPlayTriggered (#9355)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
We have been having complaints from tenants about video being in loading
screen forever. I mistakenly thought it should go to the fallback
screen, but it doesn't because fallback screen is feature-flagged only
for certain tenants. Checking the logs, it's apparent that the
`useInterval` with `play` function isn't getting called because
`onCanPlayTrigerred` was false. Since `onCanPlay` is triggered only once
and may be missed for whatever reason, it's safer to rely of
`readyState`
[docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/readyState).

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->