Add rule instance kind (#9283)
## Describe your changes
In order to prevent tenants from writing rules on both person and business, we want to prevent specific rule instances from containing instances of
both person and biz rules.

In the rule evaluation path, we'll then be able to identify the subject we are running rules on and filter our rule instances that aren't about that subject.

So, we add a `kind` to rule instance, and in the next PR we'll validate adding/updating rules so as to not mix subjects

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->