[FP-7495] Supporting copying playbooks to another tenant (#9391)
## Describe your changes
Supports copying a playbook from one tenant (to which the authed
principal has access) to another tenant (to which the authed principal
also needs to have access).

## Caveats
~I don't love that we have auth checking code in another place... it
used to be the case that to log in as another tenant, you had to hit the
`/assume_role` API to update your auth token's session to be bound to
the new tenant. This adds a new codepath....~

EDIT: I ended up doing this alternative implementation described below:

An alternative implementation could support:
- Fetching another token from the `/assume_role` API. We'd have to
update this API to be able to issue a new auth token (instead of just
statefully updating the token on the server side
- Providing the secondary token to this copy API

I'm actually thinking that I like that implementation better... it will
be some more work to build, but it would help me sleep better at night.
I think I'll still merge this, but work on that as a follow-up. Lmk what
you think :)

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->