test(e2e): FP-7538 build and run apps locally (#9376)
## Describe your changes
1. Change the E2E actions trigger back to `pull_request` from
`deployment_status`, as this was resulting in numerous action triggers
with a 'skipped' status.
2. Removed the dependency from Vercel preview URLs
3. Created a docker image to build the necessary apps for the playwright
runners.
a. The `e2e.Dockerfile` includes only the essential apps and packages
required for building. It also utilizes an alternative
version of Node and a different package manager to improve the process.
5. Created a docker image to run the tests against the other image

## Caveats
The `ubuntu-latest` machine can be very slow

## Print

![image](https://github.com/onefootprint/monorepo/assets/5114769/c1ad011c-34a5-4d2e-bb29-f6155ad395cc)