Add fn to fetch customer and save a session (#9382)
## Describe your changes
Adds a method in the incode machine code that will do all the stuff we want to approve and save a customer session.

Right now, it's in a sep thread so that it doesn't impact the main incode machine. but this introduces issues around timing if we ever want to use this in rules (probably we do).

But for now, while we're just evaluating it and looking at stuff in the dupes table, it seems okay for this to not fail/not be strictly/deterministically available when rules run (although it probably definitely will be)

## Caveats
eventually i'd like to backfill all our existing sessions, but this is a big annoying script so punting for now.

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->