Evaluate feature flag rollouts outside of LD (#9424)
## Describe your changes
Starts the work of reducing the cardinality of "keys" that we send to LaunchDarkly. It's still not clear if this is what's using more cost or if it's the number of client-side invocations

In any case - i still like this method of engaging with LD muuuuch better. We don't need LD's fancy feature flag evaluation - we literally just need them to serve us a JSON value that can be updated without deploying our server.

This switches to a new approach that evaluates whitelist/blacklist/rollout percentage inside the application instead of in launch darkly. This is a pretty reusable util that can be used for any other flags we want to migrate. I only migrate the bool flags with really high cardinality in this PR. Will migrate the JSON flags in a follow-up PR

This has another nicety: we used to hash the phone number before sending to LD because we didn't want to send PII values to LD. But since we're evaluating the flags locally now, we can remove this hashing logic and actually write our whitelist rules on specific phone numbers - much easier to test!

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Tested flag evaluations locally