[FP-7533] Backfill user_timeline seqno (#9436)
## Describe your changes
This is fun! The data lifetime table kind of acts as a timeline mapping machine-readable seqnos to human-readable timestamps. So we can use this to backfill the seqnos of user_timeline events.

This backfill looks for the seqno whose timestamp is directly after the timeline's timestamp. It then subtracts one from that seqno to get the seqno as it existed at the time that the timeline event was created.

## Caveats
 This could be brittle - if you make a UserTimeline event before saving vault data into the DB, rewinding to the time of the UserTimeline event wouldn't show it anymore... we have some coverage of this in the integration test in follow-up PRs though.

I'll need to run this with lots of manual labor. To speed up the migration, we only look for seqnos within a minute of the user timeline event. But not all timeline events have a seqno in this range. So we'll have to do multiple passes of the backfill with progressively larger (and slower) windows.

## Test Plan
Tested in some future PRs