[FP-7559] Add GET entities/<>/data (#9438)
## Describe your changes
Adds our first rewindable API - this one to fetch what vault data existed at a given seqno.

https://www.figma.com/file/5VkT8cIz0fJQsQFGaa5iN6/%F0%9F%91%A8%F0%9F%8F%BE%E2%80%8D%F0%9F%A6%B3-Users?type=design&node-id=6638-114028&mode=design&t=ngRIIdA4L5HiHUQr-4

I considered keeping this in the `GET /entities/<>` API, but i decided not to since a lot of the information in that API is not rewindable. This is a targeted API that is entirely rewindable. And another nice thing is we could even paginate this API in the future! We couldn't do that with the old `GET /entities/<>` API

## Caveats

## Test Plan
Integration test