[Coba] Short-term fix to edit vault (#9451)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

[Context on
Slack](https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1714765636645489)

#### Reason 
The edit form expects that the `formData` object will not be nested. To
ensure it, we have this function `editFormFieldName`. However, this will
(and does) fall apart when we have:
- multiple dots in `di` (we have some document dis like that)
- We have two form providers - one for decrypt and one for edit wrapped
around the components. So, when we call `useFormContext` with an intent
to populate the decrypt from, it populates the edit form as well. So,
these sanitized keys `const formField = editFormFieldName(fieldName);`
(example `NameInput`) aren't the only ones that are going to the
`formData` for edit form.

#### Solution
The PR provides a sub-optimal solution to free Coba because the actual
solution will require some clean-up. One thought I had in mind is to get
rid of `editFormFieldName` and in `convertFormData` in the `vault.tsx`,
we use a recursive helper function to get `.` separated keys from the
nested object as use them as `di` to the rest of the computation.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->