Serialize data identifiers in data_collected user timeline events (#9450)
## Describe your changes
We should move over time to have the frontend display the list of DataIdentifiers updated in a timeline event rather than the list of CDOs. This just starts to serialize the identifiers.

## Caveats

Unfortunately, there are lots of old events that don't have the list of targets that I believe still need to be backfilled... but this is also useful for integration tests, so will add anyways

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->