[FP-7570] Show more information on a document's status in the badge (#9536)
## Describe your changes
The state machine of documents got much more complex: now, a document
may be reviewed by either a human or by a machine. So, we want our
status badges to properly reflect who reviewed the document / whether
review is required.

The main objective is to make the "Custom document" situation more clear
- we put all users with custom documents into manual review, and when a
tenant goes to review a user in this state, they should see that a
document needs their review

Here are the new statuses:
- Upload failed (red) - exists today
- Upload pending (yellow) - exists today
- Uploaded manually (neutral) - this is a for a little-used case where
the tenant uploaded a driver’s license via API or via the dashboard
- Processed (neutral) - this is when we support automatic verification
of the document with one of our vendors. once the upload is complete and
the vendor has been consulted, we’ll mark the user as verified by
machine. this is identical to the current “Upload successful” state
- Review required (red) - this is a new status for custom documents that
have been uploaded by the user and need to be reviewed by a human
- Reviewed (neutral) - this is a new status for custom documents that
have been uploaded by the user and have already been reviewed by a human

A user will probably never have this many docs in this many states, but
here's to show a lot of them:

(stale screenshot, the two green badges have been updated to neutral)

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/fd1bfc98-e510-4adc-bef6-630db34e5bd0.png)


## Caveats
This runs the risk of making our document statuses much more difficult
to comprehend. Maybe we'll refine this representation over time... right
now, there's sadly just a lot of context missing.


## Test Plan
Locally, added unit tests