[1/2] Create a new junction table to support one-to-many between fingerprint <> DL (#9577)
## Describe your changes
This starts a long stack needed to support composite fingerprints.

Some background: for a while, i was thinking to just make a separate
`CompositeFingerprint` table. I got started on that, but I realized it
was going to have a lot of boilerplate columns + indexes that we've
recently fine tuned to exist on the existing `fingerprint` table. And
then, I realized there are actually very few reads o the existing
`fingerprint::lifetime_id` column, so it would be very easy to adapt the
existing fingerprint table to belong to multiple lifetimes.

The stack above just migrates to support multiple lifetimes per
fingerprint in the data model - we have a lot more work needed in order
to compute + re-compute composite fingerprints afterwards.

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->