Only show verified vaults in the dupes table (#9583)
## Describe your changes
https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1715387768750779

## Caveats

I don't like much about this code 😛 it's very brittle and requires so much context on these boolean flags on vault. I made a task to better clean up this behavior next week

We did so much work to remove the joins on the dupes fingerprint queries, but here we are again doing another join 😥

## Test Plan
Added a test that we don't see unverified users.

And, this test fails when we run it against the version of code currently deployed:
![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/56392f12-fad2-45a4-8935-f6d7f4759120.png)